+++ src/pudl/helpers.py
@@ -13,8 +13,9 @@ import json
 import pathlib
 import re
 import shutil
+import time
 from collections import defaultdict
-from collections.abc import Generator, Iterable
+from collections.abc import Callable, Generator, Iterable
 from functools import partial
 from io import BytesIO
 from typing import Any, Literal, NamedTuple
@@ -2090,3 +2091,33 @@ def check_tables_have_metadata(
     assert (
         has_no_missing_tables_with_missing_metadata
     ), f"These tables are missing datasette metadata: {tables_missing_metadata_results}"
+
+
+def retry(
+    func: Callable,
+    retry_on: tuple[type[BaseException], ...],
+    max_retries=5,
+    base_delay_sec=1,
+    **kwargs,
+):
+    """Retry a function with a short sleep between each try.
+
+    Sleeps twice as long before each retry as the last one, e.g. 1/2/4/8/16
+    seconds.
+
+    Args:
+    func: the function to retry
+    retry_on: the errors to catch.
+    base_delay_sec: how much time to sleep for the first retry.
+    kwargs: keyword arguments to pass to the wrapped function. Pass non-kwargs as kwargs too.
+    """
+    for try_count in range(max_retries):
+        delay = 2**try_count * base_delay_sec
+        try:
+            return func(**kwargs)
+        except retry_on as e:
+            logger.info(
+                f"{e}: retry in {delay}s. {try_count}/{max_retries} retries used."
+            )
+            time.sleep(delay)
+    return func(**kwargs)
